BrandIt_Back/requirements.txt
Django==3.2.14
djangorestframework==3.13.1
django-translations==1.3.0
django-translated-fields==0.12.0
psycopg2==2.9.3
django-jazzmin==2.5.0
drf-yasg==1.21.3